增加一个 ddl

+++ biz-shared/src/main/java/com/magicliang/transaction/sys/biz/shared/request/acceptance/AcceptanceCommand.java
@@ -31,7 +31,7 @@ public class AcceptanceCommand extends HandlerRequest {
      *
      * @see TransFundAccountingEntryTypeEnum
      */
-    private Short accountingEntry;
+    private Integer accountingEntry;
 
     /**
      * 支付备注，上游系统选填
